RIA runs all work: different Spinup() result for presentDay
RIApresentDayRuns has a different age raster (2015) then the three other runs (FRI, harvest1 and harvest2 all have 2020). Because inventory ages are a unique identifier for the pixel groups, the pixelGroups are different for presentDayRuns (in level3DT) and the spinupResults are different then for the three other RIA runs which all have the same spinupResults.